refactor(durable-streams): merge server + tailer into single observer with colors
Combine demo/server.ts and demo/tail.ts into demo/observe.ts — a single
process that starts the durable streams server and tails the journal via
SSE with ANSI color-coded output (cyan for yields, magenta for closes,
green/red/yellow for ok/err/cancelled status).

Simplify tmux layout from 4 panes to 3:
- Observer (server + journal tail)
- Cook workflow (focused)
- Control (kill command)
