+++ durable-streams/demo/README.md
@@ -23,31 +23,27 @@ Or with a custom stream ID:
 ./demo/start.sh my-stream
 ```
 
-This opens a `tmux` session named `durable-dinner` with 4 panes:
+This opens a `tmux` session named `durable-dinner` with 3 panes:
 
-- Server (`demo/server.ts`)
-- Journal tailer (`demo/tail.ts`)
-- Cook workflow (`demo/cook.ts`) (focused)
-- Control pane with pre-typed kill command
+```
+┌────────────────────────────────────┐
+│  Observer (server + journal tail)  │
+├───────────────────┬────────────────┤
+│  Cook (focused)   │  Control       │
+└───────────────────┴────────────────┘
+```
 
-Press Enter in the cook pane to start.
+- **Observer** (`demo/observe.ts`) — starts the server and tails the journal
+  via SSE, printing color-coded events as they arrive
+- **Cook** (`demo/cook.ts`) — the durable cooking workflow (focused, press Enter)
+- **Control** — pre-typed kill command to simulate a crash
 
 ## Demo Script
 
 1. Start with `./demo/start.sh`
-2. In the cook pane, press Enter to run:
-
-   ```sh
-   node --experimental-strip-types demo/cook.ts
-   ```
-
-3. Let it run a few steps
-4. In the control pane, press Enter to hard-kill:
-
-   ```sh
-   bash -lc 'PGID=$(ps -o pgid= -p $(tmux display-message -p -t "durable-dinner:0.2" "#{pane_pid}") | tr -d " " ); kill -9 -$PGID'
-   ```
-
+2. In the cook pane, press Enter to run the workflow
+3. Watch color-coded journal events stream in the observer pane
+4. In the control pane, press Enter to hard-kill the cook process
 5. Back in the cook pane, rerun:
 
    ```sh
@@ -57,34 +53,27 @@ Press Enter in the cook pane to start.
 You should see:
 
 - `Found N events in journal — replaying...`
-- No new tailer events during replay
+- No new observer events during replay
 - New events only after replay catches up and live execution resumes
 
 ## Run Without tmux
 
-Open 3 terminals from `durable-streams/`:
+Open 2 terminals from `durable-streams/`:
 
-Terminal 1 (server):
-
-```sh
-node --experimental-strip-types demo/server.ts
-```
-
-Terminal 2 (tailer):
+Terminal 1 (observer — server + tailer):
 
 ```sh
-DURABLE_STREAM_ID=my-stream node --experimental-strip-types demo/tail.ts
+DURABLE_STREAM_ID=my-stream node --experimental-strip-types demo/observe.ts
 ```
 
-Terminal 3 (workflow):
+Terminal 2 (workflow):
 
 ```sh
 DURABLE_STREAM_ID=my-stream node --experimental-strip-types demo/cook.ts
 ```
 
-Then kill and restart Terminal 3 to observe replay behavior.
+Then kill and restart Terminal 2 to observe replay behavior.
 
 ## Environment Variables
 
-- `DURABLE_SERVER_URL` (default: `http://localhost:4437`)
 - `DURABLE_STREAM_ID` (default: `dinner-demo`)
