+++ durable-streams/demo/start.sh
@@ -2,15 +2,15 @@
 #
 # Durable Dinner — tmux demo launcher
 #
-# Starts a 4-pane tmux session:
+# Starts a 3-pane tmux session:
 #
-#   ┌──────────────────┬──────────────────┐
-#   │  Server           │  Journal Tailer  │  30%
-#   │  (demo:server)    │  (demo:tail)     │
-#   ├──────────────────┬──────────────────┤
-#   │  Cook (focused)   │  Control         │  70%
-#   │  (demo:cook)      │  (kill command)  │
-#   └──────────────────┴──────────────────┘
+#   ┌────────────────────────────────────┐
+#   │  Observer (server + journal tail)  │  40%
+#   │  (demo:observe)                    │
+#   ├───────────────────┬────────────────┤
+#   │  Cook (focused)   │  Control       │  60%
+#   │  (demo:cook)      │  (kill cmd)    │
+#   └───────────────────┴────────────────┘
 #
 # Usage:
 #   ./demo/start.sh [stream-id]   # launch with optional stream ID
@@ -26,7 +26,7 @@ DIR="$(cd "$(dirname "$0")/.." && pwd)"
 STREAM_ID="${1:-dinner-demo}"
 NODE="node --experimental-strip-types"
 
-# Export so all panes inherit it (cook.ts and tail.ts read from env)
+# Export so all panes inherit it
 export DURABLE_STREAM_ID="$STREAM_ID"
 
 # ------------------------------------------------------------------
@@ -50,53 +50,45 @@ tmux kill-session -t "$SESSION" 2>/dev/null || true
 # Build the layout
 # ------------------------------------------------------------------
 
-# Pane 0 (top-left): Durable Streams server
+# Pane 0 (top): Observer (server + journal tailer)
 tmux new-session -d -s "$SESSION" -c "$DIR" -x 200 -y 50
 
-# Set stream ID as a tmux environment variable — all new panes/shells
-# in this session inherit it automatically. Avoids inline quoting issues.
+# Set stream ID as a tmux environment variable
 tmux set-environment -t "$SESSION" DURABLE_STREAM_ID "$STREAM_ID"
 
-# Split horizontally — bottom gets 70%, top gets 40%
+# Split horizontally — bottom gets 60% (cook + control), top gets 40% (observer)
 tmux split-window -v -t "$SESSION" -c "$DIR" -p 60
 
-# Split the top pane vertically — right side becomes tailer
-tmux split-window -h -t "${SESSION}:0.0" -c "$DIR" -p 50
-
 # Split the bottom pane vertically — right side becomes control
-tmux split-window -h -t "${SESSION}:0.2" -c "$DIR" -p 50
+tmux split-window -h -t "${SESSION}:0.1" -c "$DIR" -p 40
 
 # After all splits, pane indices are:
-#   0 = top-left     (Server)
-#   1 = top-right    (Tailer)
-#   2 = bottom-left  (Cook)
-#   3 = bottom-right (Control)
+#   0 = top          (Observer)
+#   1 = bottom-left  (Cook)
+#   2 = bottom-right (Control)
 
 # ------------------------------------------------------------------
 # Start processes
 # ------------------------------------------------------------------
 
-# Pane 0: Start the server
-tmux send-keys -t "${SESSION}:0.0" "$NODE demo/server.ts" Enter
+# Pane 0: Start the observer (server + tailer)
+tmux send-keys -t "${SESSION}:0.0" "$NODE demo/observe.ts" Enter
 
 # Give the server a moment to bind its port
 sleep 2
 
-# Pane 1: Start the journal tailer
-tmux send-keys -t "${SESSION}:0.1" "$NODE demo/tail.ts" Enter
+# Pane 1: Pre-type the cook command (presenter hits Enter when ready)
+tmux send-keys -t "${SESSION}:0.1" "$NODE demo/cook.ts"
 
-# Pane 2: Pre-type the cook command (presenter hits Enter when ready)
-tmux send-keys -t "${SESSION}:0.2" "$NODE demo/cook.ts"
-
-# Pane 3: Pre-type a pane-scoped kill command for the cook pane process group
-tmux send-keys -t "${SESSION}:0.3" "bash -lc 'PGID=\$(ps -o pgid= -p \$(tmux display-message -p -t \"${SESSION}:0.2\" \"#{pane_pid}\") | tr -d \" \" ); kill -9 -\$PGID'"
+# Pane 2: Pre-type a pane-scoped kill command for the cook pane process group
+tmux send-keys -t "${SESSION}:0.2" "bash -lc 'PGID=\$(ps -o pgid= -p \$(tmux display-message -p -t \"${SESSION}:0.1\" \"#{pane_pid}\") | tr -d \" \" ); kill -9 -\$PGID'"
 
 # ------------------------------------------------------------------
 # Focus & attach
 # ------------------------------------------------------------------
 
 # Focus the cook pane (bottom-left)
-tmux select-pane -t "${SESSION}:0.2"
+tmux select-pane -t "${SESSION}:0.1"
 
 # Attach (or switch if already inside tmux)
 if [ -n "${TMUX:-}" ]; then
